refactor: move row id metadata into submodule
